Assert that route path must be a String
This validates the route path input to make sure it's a String, or an Array with a String as the first item.

The path-to-regexp lib supports String, RegExp, or an Array of either as its input for `path`. The idea here is to make this package have a similar signature, but this package will throw when it encounters a RegExp route because it cannot produce a deterministic string from it.